Deleted class consumableItem
Class consumableItem does not properly extend class genericItem as can be seen by the duplicate code within it. If code in this class is 
modified and the duplicate code is not removed, 'shadowing' will likely occur resulting in problems in object creation. 

Methods for item recovery may not work as intended in that supplying the method parameters may be difficult to perform at a per 
item basis. Methods also recover more than what they are stated to recover (sum of user hp and item recovery points are returned).
It would be best to have this file removed for they do not adequately suit where the project is going...at least for now.